Recognition threshold moved to int (may need clear data to remove old double value).

+++ src/com/eim/utilities/EIMPreferences.java
@@ -117,8 +117,8 @@ public class EIMPreferences {
 	private String recognitionThresholdKey;
 	private String recognitionThresholdDefault;
 
-	public double recognitionThreshold() {
-		return Double.parseDouble(mSharedPreferences.getString(
+	public int recognitionThreshold() {
+		return Integer.parseInt(mSharedPreferences.getString(
 				recognitionThresholdKey, recognitionThresholdDefault));
 	}
